open_banking_api_mock/requirements.txt
fastapi==0.95.2         # Основной фреймворк для создания API
uvicorn==0.22.0        # ASGI-сервер для запуска FastAPI
python-jose==3.3.0     # Библиотека для работы с JWT (JSON Web Tokens)
passlib==1.7.4         # Библиотека для хэширования паролей
python-dotenv==0.20.0  # Для работы с переменными окружения из .env файлов